+++ library/termios_headers.h
@@ -48,8 +48,8 @@
 
 /****************************************************************************/
 
-extern int __termios_console_hook(struct fd *,struct file_action_message *);
-extern int __termios_serial_hook(struct fd *,struct file_action_message *);
+extern int64_t __termios_console_hook(struct fd *,struct file_action_message *);
+extern int64_t __termios_serial_hook(struct fd *,struct file_action_message *);
 extern struct termios *__get_termios(struct fd *);
 extern int openserial(const char *device_name,int unit,int mode);
